split up document generator to link and page generator
